refactor(06-02): extract BracketManager (assembly + fill-anchored children)
- Move _assemble_bracket_and_emit + _create_fill_anchored_children verbatim
  (TAB, no queue) to brackets/bracket_manager.py (D-04/D-08/D-13)
- BracketManager injects order_storage, logger, coordinator-owned BracketBook (D-09)
- OrderManager constructs it once in __init__; 3 call sites delegate (D-04 star)
- brackets/__init__ re-exports BracketManager + BracketBook
- Remove now-dead imports (assert_never, PercentFromDecision/FromFill, _PendingBracket)
- Golden byte-exact (134 / 46189.87730727451); e2e 58/58; mypy strict clean
- order_handler.py + order_handler/__init__.py byte-unchanged
