+++ itrader/order_handler/brackets/__init__.py
@@ -1,12 +1,13 @@
 """
 Brackets subdomain package.
 
-Re-exports the BracketBook pending-bracket state owner (D-04/D-05) so consumer
-import paths stay short after the order-manager decomposition (pure move,
-D-12/D-13). _PendingBracket stays internal (leading-underscore). The
-BracketManager joins this barrel in plan 02.
+Re-exports the BracketManager + the BracketBook pending-bracket state owner
+(D-04/D-05) so consumer import paths stay short after the order-manager
+decomposition (pure move, D-12/D-13). The stateless levels helper and
+_PendingBracket stay internal (leading-underscore).
 """
 
+from .bracket_manager import BracketManager
 from .bracket_book import BracketBook
 
-__all__ = ["BracketBook"]
+__all__ = ["BracketManager", "BracketBook"]
